refactor: Move `asyncsync.py` to src root directory
